fix: importDir uses the same structure as source directory
Fixes #84

The files were uploaded with parent folder's id as path. So they were
not in the directory they belong to, but in a directory located at the
root and named after the real file's parent directory id.

This is fixed by  giving the parent folder's path instead of id to the
`uploadFile` function.